requirements.txt
aiocryptopay==0.3.6
aiohttp==3.9.3
aiomysql==0.2.0
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
antiorm==1.2.1
anyio==4.3.0
APScheduler==3.10.4
asyncio==3.4.3
attrs==23.2.0
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
cards==1.0.1
certifi==2023.11.17
charset-normalizer==2.0.12
cleo==2.1.0
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
crashtest==0.4.1
db==0.1.1
db-sqlite3==0.0.1
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
env==0.1.0
et-xmlfile==1.1.0
fastjsonschema==2.19.1
filelock==3.13.4
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.3
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
mysql-connector-python==8.3.0
numpy==1.26.4
openai==1.14.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
peewee==3.17.1
peewee-async==0.10.0
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.12.0
PyMySQL==1.1.0
pyparsing==3.1.2
pyproject_hooks==1.0.0
pyTelegramBotAPI==4.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
rapidfuzz==3.8.1
requests==2.27.1
requests-toolbelt==1.0.0
rich==10.7.0
rsa==4.9
schedule==1.2.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
StrEnum==0.4.15
tinydb==4.5.1
tomlkit==0.12.4
tqdm==4.66.2
trove-classifiers==2024.4.10
typer==0.3.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.25.3
yarl==1.9.4
zipp==3.18.1
